aula_virtual/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-restframework==0.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
